review rename beforeAddUserToARoom - > beforeAddUserToRoom
