+++ apps/meteor/lib/callbacks/beforeAddUserToRoom.ts
